[fix] RC5 for XKaapi: - warning when compiling on 32/64 bits machines - type 64 -> 32 - mac os x and Linux - move kaapi_accesscw_t in kaapi.h - partitionning as default option in configure. - missing header stdint.h - move kaapi_accesscw_t in kaapi.h - change format definition with const char* in place of std::stirng to avoid warning with gcc > 4.4 - update bug in execframe introduced by previous bad merge - update Athapascan API due to change in task format. - add horner example  (C and C++) exposes the future c++ api for linear work adaptive processing - add variance example

+++ src/machine/common/kaapi_hwtopo.c
@@ -209,7 +209,7 @@ int kaapi_hw_init(void)
         ncpu = hwloc_cpuset_weight( obj->cpuset );
         kaapi_default_param.memory.levels[memdepth].count = 1;
         kaapi_default_param.memory.levels[memdepth].affinity = (kaapi_affinityset_t*)calloc( 1, sizeof(kaapi_affinityset_t) );
-        kaapi_default_param.memory.levels[memdepth].affinity[0].mem_size = obj->memory.total_memory;
+        kaapi_default_param.memory.levels[memdepth].affinity[0].mem_size = (size_t)obj->memory.total_memory;
         kaapi_default_param.memory.levels[memdepth].affinity[0].ncpu = ncpu = hwloc_cpuset_weight( obj->cpuset );
         kaapi_default_param.memory.levels[memdepth].affinity[0].type = KAAPI_MEM_NODE;
         kaapi_hwcpuset2affinity(
@@ -231,9 +231,9 @@ int kaapi_hw_init(void)
       while (obj !=0)
       {
         if (obj->type == HWLOC_OBJ_NODE)
-          kaapi_default_param.memory.levels[memdepth].affinity[idx].mem_size = obj->memory.local_memory;
+          kaapi_default_param.memory.levels[memdepth].affinity[idx].mem_size = (size_t)obj->memory.local_memory;
         else
-          kaapi_default_param.memory.levels[memdepth].affinity[idx].mem_size = obj->attr->cache.size;
+          kaapi_default_param.memory.levels[memdepth].affinity[idx].mem_size = (size_t)obj->attr->cache.size;
 
         kaapi_default_param.memory.levels[memdepth].affinity[idx].ncpu = ncpu = hwloc_cpuset_weight( obj->cpuset );
         kaapi_default_param.memory.levels[memdepth].affinity[idx].type = (obj->type == HWLOC_OBJ_CACHE ? KAAPI_MEM_CACHE : KAAPI_MEM_NODE);
